pyproject.toml
[project]
name = "python-whatsapp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=26.1.0",
    "dotenv>=0.9.9",
    "fastapi>=0.128.1",
    "google-auth>=2.48.0",
    "google-cloud-firestore>=2.23.0",
    "httpx>=0.28.1",
    "isort>=7.0.0",
    "openai>=2.16.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "uvicorn>=0.40.0",
]
